backend/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.6.0
pydantic==2.10.6
psycopg[binary]==3.2.6
celery[redis]==5.4.0
redis==5.2.1
PyYAML==6.0.2
pytest==8.3.5
pytest-django==4.11.1
factory_boy==3.3.3
